feat(api): Introduce new core unit tests and thread-safety guard
This commit introduces a comprehensive new test suite (`cdd_file_tests.cpp`) that directly exercises the `DataReader` and `DataWriter` APIs. These tests verify core functionality including file creation, chunk appending (raw and compressed), multi-block index chaining, appending to existing files, and in-memory backends.

During the implementation of these tests, a critical bug was discovered and fixed: callers were incorrectly labeling raw data as `ZSTD_COMPRESSED` without performing compression. The new tests now correctly compress data before appending it, ensuring the ZSTD read path is properly validated.

Major changes and fixes include:

- **New Test Suite:** Added `cdd_file_tests.cpp` for direct, low-level validation of the core file I/O components.
- **Fix (Tests):** Corrected all new tests to manually compress data with `ZstdCompressor` before passing it to `DataWriter` with the `ZSTD_COMPRESSED` type. This resolves all "Invalid ZSTD frame" errors.
- **Refactor (DataWriter):** The `append_chunk` API has been improved to use move semantics for the data payload, avoiding unnecessary copies. This was accomplished by changing the signature to accept a `Chunk&` and implementing an internal RAII `ChunkDataGuard` for safe, exception-proof data transfer.
- **Refactor (Hashing):** The `calculate_blake3_hash128` function is now properly overloaded to transparently handle `std::vector`, `memory::vector`, and `std::span`.
- **Feat (Concurrency):** Introduced a `ConcurrencyGuard` in `CddContext` to detect and prevent concurrent use of a single context handle, which is not thread-safe. This improves robustness by failing fast in multi-threaded scenarios.
- **Fix (Build):** Resolved a runtime issue on Windows by correctly linking `mimalloc` to the test executable, ensuring the custom allocator is available at runtime in debug builds. The CMakeLists were cleaned up to handle this more robustly.

+++ src/c_api/operations/operation_types.h
@@ -63,7 +63,7 @@ struct ChunkWriteDetails {
     size_t chunk_index;
     int64_t original_size;
     int64_t compressed_size;
-    double compression_ratio;
+    float compression_ratio;
 };
 
 struct StoreChunkResponse : OperationResponseBase {
